Move classes into sensible packages
