--- src/TaxSvc/CancelTaxResponse.java
@@ -1,11 +1,0 @@
-package TaxSvc;
-
-import TaxSvc.CommonResponse.Message;
-import TaxSvc.CommonResponse.SeverityLevel;
-
-public class CancelTaxResponse {
-	public CancelTaxResult CancelTaxResult;
-	public SeverityLevel ResultCode;
-	public Message[] Messages;
-	
-}
