+++ src/com/avatax/tax/GetTaxResult.java
@@ -1,4 +1,4 @@
-package TaxSvc;
+package com.avatax.tax;
 
 import java.util.Date;
