+++ src/com/avatax/tax/TaxOverrideDef.java
@@ -1,4 +1,4 @@
-package TaxSvc;
+package com.avatax.tax;
 
 public class TaxOverrideDef
 {
