+++ src/com/avatax/test/PingTest.java
@@ -1,5 +1,9 @@
-import TaxSvc.*;
-import TaxSvc.CommonResponse.*;
+package com.avatax.test;
+
+import com.avatax.tax.CommonResponse.*;
+import com.avatax.tax.GeoTaxResult;
+import com.avatax.tax.TaxSvc;
+
 import java.io.*;
 import java.util.Properties;
